refactor : manage 모달 렌더 최적화 이동
