--- src/1_Page/Team/ui/TeamManageButtonGroupProps/type.d.ts
@@ -1,3 +1,0 @@
-type TeamManageButtonGroupProps = {
-  handleToggleManageModal: () => void;
-};
