+++ src/1_Page/Team/ui/TeamManagePanel/type.d.ts
@@ -1,0 +1,4 @@
+type TeamManageButtonGroupProps = {
+  teamInfo: TeamInfo;
+  handlers: ManageHandlers;
+};
