+++ src/1_Page/Team/ui/TeamManagePanel/ui/ManageModal/ui/AutoMatchPanel/type.d.ts
