+++ src/1_Page/Team/ui/TeamManagePanel/ui/ManageModal/ui/BannerImageInput/model/type.d.ts
