+++ src/1_Page/Team/ui/TeamManagePanel/ui/ManageModal/ui/BannerImageInput/type.d.ts
