+++ src/1_Page/Team/ui/TeamManagePanel/ui/ManageModal/ui/EmblemImageInput/type.d.ts
