+++ src/1_Page/Team/ui/TeamManagePanel/ui/ManageModal/ui/TeamApplications/type.d.ts
