+++ src/1_Page/Team/ui/TeamManagePanel/ui/ManageModal/ui/TeamApplications/ui/TeamApplicationItem/index.tsx
@@ -1,7 +1,7 @@
-import useDeleteApproveMember from "../../../../../../../../3_Entity/Team/useDeleteApproveMember";
-import usePostApproveMember from "../../../../../../../../3_Entity/Team/usePostApproveMember";
-import { platform } from "../../../../../../../../4_Shared/constant/platform";
-import { getPlatformIcon } from "../../../../../../../../4_Shared/lib/getPlatformIcon";
+import useDeleteApproveMember from "../../../../../../../../../../3_Entity/Team/useDeleteApproveMember";
+import usePostApproveMember from "../../../../../../../../../../3_Entity/Team/usePostApproveMember";
+import { platform } from "../../../../../../../../../../4_Shared/constant/platform";
+import { getPlatformIcon } from "../../../../../../../../../../4_Shared/lib/getPlatformIcon";
 
 const TeamApplicationItem = (props: TeamApplicationItemProps) => {
   const { player, team_list_idx, addDisplayPlayer } = props;
