+++ src/1_Page/Team/ui/TeamManagePanel/ui/ManageModal/ui/TextInputForm/lib/schema.ts
