+++ src/1_Page/Team/ui/TeamManagePanel/ui/ManageModal/ui/TextInputForm/ui/TeamNameCheckInput/model/type.d.ts
