+++ src/1_Page/Team/ui/TeamManagePanel/ui/ManageModal/ui/TextInputForm/ui/TeamNameCheckInput/type.d.ts
